Fixed fucked up CL namespace
Provided skeleton had no namespace references, and compilation namespaces aren't even configured correctly.

Absolute aids.
