+++ TestingStaff/TestingStaff.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{A3AA37D5-DAA0-4AF0-AD82-419646520965}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Testing4</RootNamespace>
-    <AssemblyName>Testing4</AssemblyName>
+    <RootNamespace>TestingStaff</RootNamespace>
+    <AssemblyName>TestingStaff</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
@@ -50,11 +50,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="tstCustomer.cs" />
-    <Compile Include="tstOrder.cs" />
     <Compile Include="tstStaff.cs" />
-    <Compile Include="tstStock.cs" />
-    <Compile Include="UnitTest1.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
